requirements.txt
# Python AI开发框架依赖

# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
fastapi-pagination==0.12.14  # 分页插件

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0
PyJWT>=2.8.0

# 异步任务
celery==5.3.4
kombu==5.3.4

# 缓存
redis==5.0.1
pymemcache==4.0.0

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# AI相关
torch==2.1.1
transformers==4.36.0
onnx==1.15.0
mlflow==2.8.1
scikit-learn==1.3.2
numpy==1.24.4
pandas>=2.0.0

# 数据处理
openpyxl>=3.1.0

# 文件处理
python-magic==0.4.27
Pillow>=10.0.0

# 网络请求
requests>=2.31.0
aiohttp>=3.8.0  # 异步HTTP客户端，用于TTS API调用

# 系统监控
prometheus-client==0.19.0
psutil>=5.9.0

# 时间处理
python-dateutil>=2.8.2
pytz>=2023.3

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# 部署
gunicorn==21.2.0
docker==6.1.3